+++ src/bot.py
@@ -1,9 +1,8 @@
-from telegram.ext import Application, ApplicationBuilder, CallbackQueryHandler, PicklePersistence
+from telegram.ext import Application, ApplicationBuilder, PicklePersistence
 
 from conversation import start_conversation
 from core import config
 from jobs import monthly_bill_reminder_job, monthly_stat_job, weekly_stat_job
-from service.callback_router import callback_router
 
 
 def create_bot():
@@ -19,7 +18,6 @@ def create_bot():
         monthly_bill_reminder_job, when=config.MONTHLY_RECEIPT_REMINDER_TIME, day=config.MONTHLY_RECEIPT_REMINDER_DAY
     )
     bot_app.job_queue.run_monthly(monthly_stat_job, when=config.MONTHLY_STAT_TIME, day=config.MONTHLY_STAT_DAY)
-    bot_app.add_handler(CallbackQueryHandler(callback_router))
     return bot_app
