+++ src/constants/callback_data.py
@@ -8,5 +8,5 @@ CALLBACK_IS_EXPERT_COMMAND = "CALLBACK_IS_EXPERT_COMMAND"
 CALLBACK_NOT_EXPERT_COMMAND = "CALLBACK_NOT_EXPERT_COMMAND"
 CALLBACK_REGISTR_AS_EXPERT_COMMAND = "CALLBACK_REGISTR_AS_EXPERT_COMMAND"
 CALLBACK_SUPPORT_OR_CONSULT_COMMAND = "CALLBACK_SUPPORT_OR_CONSULT_COMMAND"
-CALLBACK_DONE_BILL_COMMAND = "callback_done_bill_command"
-CALLBACK_SKIP_BILL_COMMAND = "callback_skip_bill_command"
+CALLBACK_DONE_BILL_COMMAND = "CALLBACK_DONE_BILL_COMMAND"
+CALLBACK_SKIP_BILL_COMMAND = "CALLBACK_SKIP_BILL_COMMAND"
