+++ src/jobs.py
@@ -1,11 +1,14 @@
 from datetime import timedelta
 from string import Template
 
+from telegram import InlineKeyboardButton, InlineKeyboardMarkup
 from telegram.ext import CallbackContext
 
-from core.send_message import send_statistics
+from constants.callback_data import CALLBACK_DONE_BILL_COMMAND, CALLBACK_SKIP_BILL_COMMAND
+from core import config
+from core.send_message import send_message, send_statistics
 from service.api_client import APIService
-from service.bill import daily_bill_remind_job
+from service.repeat_message import repeat_after_one_hour_button
 
 
 async def weekly_stat_job(context: CallbackContext) -> None:
@@ -77,3 +80,30 @@ async def monthly_bill_reminder_job(context: CallbackContext) -> None:
     user_list = bill_stat.telegram_ids
     for telegram_id in user_list:
         context.job_queue.run_once(daily_bill_remind_job, when=timedelta(seconds=1), user_id=telegram_id)
+
+
+async def daily_bill_remind_job(context: CallbackContext) -> None:
+    """
+    Send message every day until delete job from JobQueue
+    :param context:
+    :return:
+    """
+    job = context.job
+    message = "Вам необходимо сформировать чек"
+    bill_done_button = InlineKeyboardButton(text="✅ Уже отправил(а)", callback_data=CALLBACK_DONE_BILL_COMMAND)
+    bill_skip_button = InlineKeyboardButton(text="🕑 Скоро отправлю", callback_data=CALLBACK_SKIP_BILL_COMMAND)
+    menu = InlineKeyboardMarkup([[repeat_after_one_hour_button], [bill_done_button], [bill_skip_button]])
+    await send_message(chat_id=job.user_id, text=message, reply_markup=menu, context=context)
+    send_time = config.MONTHLY_RECEIPT_REMINDER_TIME
+    # user_utc = context.user_data.get("UTC")
+    # Не смог понять, в каком виде хранятся данные о часовом поясе юзера. Здесь надо переопределить информацию о
+    # времени отправки сообщения
+    # if user_utc:
+    #     send_time += user_utc
+
+    context.job_queue.run_daily(
+        daily_bill_remind_job,
+        time=send_time,
+        user_id=job.user_id,
+        name=f"send_{job.user_id}_bill_until_complete",
+    )
